allow deployment only on main branch

+++ scripts/deploy.sh
@@ -2,10 +2,14 @@
 
 set -e
 
+BRANCH="$(git rev-parse --abbrev-ref HEAD)"
+if [ "$BRANCH" != "main" ]; then
+    echo 'You are on the wrong branch! Switch to main branch and try again.';
+    exit 1;
+fi
+
 rm -rf dist
-
 npm run build
-
 cd dist
 
 git init
